+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changes
 
+* [FEATURE][rust] New `miden-web3signer-authenticator` crate provides `Web3SignerAuthenticator`, which allows a Miden client to sign transactions with `EcdsaK256Keccak` keys held in a Web3Signer instance ([#2501](https://github.com/0xMiden/rust-sdk/pull/2501)).
 * Updated workspace version to 0.16.1 ([#2526](https://github.com/0xMiden/rust-sdk/pull/2526)).
 
 ### Fixes